major breakthrough using ThreadLocals
class instance variables are not possible to use as they reference the
first instance created of which the DSL was parsed and cached. Caching
is mandatory as it is 20000x faster than creating the DSL for each
instance. With ThreadLocals the instance members can be emulated via
the companion object, they are set and reset in handle().